pyproject.toml
[project]
name = "smolagents-demo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "datasets>=3.2.0",
    "googlemaps>=4.10.0",
    "huggingface-hub>=0.27.1",
    "langchain>=0.3.17",
    "langchain-community>=0.3.16",
    "marimo>=0.11.5",
    "pandas>=2.2.3",
    "python-dotenv>=1.0.1",
    "rank-bm25>=0.2.2",
    "sentence-transformers>=3.4.1",
    "smolagents[litellm,transformers]>=1.5.0",
    "sqlalchemy>=2.0.37",
]
